refactor(checker): split circular computed helpers
